+++ documentation/code/index.ts
@@ -1,6 +1,6 @@
 export {BarChartDemo} from './BarChartDemo';
 export {GroupedBarChartDemo} from './GroupedBarChartDemo';
 export {LineChartDemo} from './LineChartDemo';
-export {NormalizedStackedBarDemo} from './NormalizedStackedBarDemo';
+export {NormalizedStackedBarChartDemo} from './NormalizedStackedBarChartDemo';
 export {SparklineDemo} from './SparklineDemo';
 export {StackedAreaChartDemo} from './StackedAreaChartDemo';