+++ src/components/NormalizedStackedBar/index.ts
@@ -1,2 +1,0 @@
-export {NormalizedStackedBar} from './NormalizedStackedBar';
-export {Orientation, Size} from './types';