+++ src/components/NormalizedStackedBarChart/index.ts
@@ -1,0 +1,2 @@
+export {NormalizedStackedBarChart} from './NormalizedStackedBarChart';
+export {Orientation, Size} from './types';